// src/declarations/cosmere-rpg/types/config.d.ts
export interface AttributeGroupConfig {
    key: string;
    label: string;
    attributes: [Attribute, Attribute];
    resource: Resource;
}